// src/app/model/user.ts
export class User {
  firstName: string;
  lastName: string;
  userTypeId: number;
  branchId:number;
  branchDb: string;
  username: string;
  refreshToken: string;
  photoURL: string;
  emailVerified: number;
  mobileVerified: number;
  roleId: number;
  userId: string;
  userType: string;
  userTypeSpeId: number;
  upDatetime: string;
  uid: string;
  mobileNo: any;
  aboutUs: string;
  accessToken: string;
  active: string;
  designation: string;
  displayName: string;
  email: string;
  location: any;
  organisation: string;
  mode: number;
  directIconId : number;
  logoUrl: any;
  constructor() {
  }


}